app/requirements.txt
# requirements.txt
fastapi
cassandra-driver
uvicorn
pydantic
